add jwt,repo,usecase

+++ internal/models/note.go
@@ -3,8 +3,8 @@ package models
 import "time"
 
 type Note struct {
-	ID        int64     `json:"id"`
+	ID        int       `json:"id"`
 	Content   string    `json:"content"`
-	UserID    int64     `json:"userId"`
+	UserID    int       `json:"userId"`
 	CreatedAt time.Time `json:"createdAt"`
 }
